move fns out of utils
